+++ admin/client/src/components/HtmlEditorField/HtmlEditorField.js
@@ -1,79 +1,33 @@
-import React from 'react';
-import SilverStripeComponent from 'lib/SilverStripeComponent';
+import React, { PropTypes, Component } from 'react';
 import fieldHolder from 'components/FieldHolder/FieldHolder';
 import ReactTinyMCE from 'react-tinymce';
-import TinyMCEInsertMediaPlugin from './TinyMCEInsertMediaPlugin';
+import tinymceLoader from './tinymceLoader';
+import insertMediaPlugin from './insertMediaPlugin';
 
-/**
- * NOTE: Paste plugin doesn't not always work
- * https://github.com/tinymce/tinymce/issues/2728
- */
-
-class HtmlEditorField extends SilverStripeComponent {
+class HtmlEditorField extends Component {
 
   constructor(props) {
     super(props);
 
     this.handleChange = this.handleChange.bind(this);
-    this.checkLoadedLibrary = this.checkLoadedLibrary.bind(this);
-    this.loadLibrary = this.loadLibrary.bind(this);
     this.getEditor = this.getEditor.bind(this);
-
-    this.state = {
-      loaded: false,
-    };
-    if (!window.tinymce) {
-      this.loadLibrary();
-    } else {
-      this.state.loaded = true;
-    }
   }
 
   getEditor() {
     return window.tinymce.EditorManager.get(this.props.id);
   }
 
-  loadLibrary() {
-    const id = 'react-tinymce-loader';
-    if (!document.getElementById(id)) {
-      // lazy loading TinyMCE this way (or easger loading through page template),
-      // otherwise bundling TinyMCE causes more problems than it's worth.
-      const script = document.createElement('script');
-      script.type = 'text/javascript';
-      script.id = 'react-tinymce-loader';
-      script.async = true;
-      script.src = `${this.props.data.config.baseURL}/tinymce.min.js`;
-      document.getElementsByTagName('head')[0].appendChild(script);
-    }
-    this.checkLoadedLibrary();
-  }
-
-  checkLoadedLibrary() {
-    if (window.tinymce) {
-      this.setState({ loaded: true });
-    } else {
-      setTimeout(this.checkLoadedLibrary, 100);
-    }
-  }
-
   render() {
-    if (!this.state.loaded) {
-      return null;
-    }
-
     return (
-      <div className="htmleditor-wrapper">
-        <ReactTinyMCE
-          id={this.props.id}
-          name={this.props.name}
-          content={this.props.value}
-          config={this.props.data.config}
-          onChange={this.handleChange}
-          onUndo={this.handleChange}
-          onRedo={this.handleChange}
-        />
-        <TinyMCEInsertMediaPlugin editorId={this.props.id} />
-      </div>
+      <ReactTinyMCE
+        id={this.props.id}
+        name={this.props.name}
+        content={this.props.value}
+        config={this.props.data.config}
+        onChange={this.handleChange}
+        onUndo={this.handleChange}
+        onRedo={this.handleChange}
+      />
     );
   }
 
@@ -91,15 +45,17 @@ class HtmlEditorField extends SilverStripeComponent {
 }
 
 HtmlEditorField.propTypes = {
-  extraClass: React.PropTypes.string,
-  id: React.PropTypes.string.isRequired,
-  name: React.PropTypes.string.isRequired,
-  onChange: React.PropTypes.func,
-  value: React.PropTypes.oneOfType([React.PropTypes.string, React.PropTypes.number]),
-  readOnly: React.PropTypes.bool,
-  disabled: React.PropTypes.bool,
-  placeholder: React.PropTypes.string,
-  type: React.PropTypes.string,
+  extraClass: PropTypes.string,
+  id: PropTypes.string.isRequired,
+  name: PropTypes.string.isRequired,
+  onChange: PropTypes.func,
+  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  readOnly: PropTypes.bool,
+  disabled: PropTypes.bool,
+  placeholder: PropTypes.string,
+  data: PropTypes.shape({
+    config: PropTypes.object,
+  }),
 };
 
 HtmlEditorField.defaultProps = {
@@ -109,6 +65,6 @@ HtmlEditorField.defaultProps = {
   className: '',
 };
 
-export { HtmlEditorField };
+export { HtmlEditorField, insertMediaPlugin, tinymceLoader };
 
-export default fieldHolder(HtmlEditorField);
+export default fieldHolder(tinymceLoader(insertMediaPlugin(HtmlEditorField)));
